Unpin funowl and fix bug in test example.
The compliance test had a data bug where a linkml class intended as
anonymous was being mapped to an owl Class declaration, rather than
anon expression.

This bug was masked (harmlessly) in previous versions of funowl,
but in newer versions it attempted to write a declaration with null
uri and raised error.

Fixed example to have the right declaration; also made the dumper
code more defensive, and throw more meaningful error in case of
declarations that can't be fulfilled.

Fixes #32

This PR also unpins the previously defensive pinning of funowl,
allowing funowl 0.2.x and thus newer versions of rdflib,
replaces/fixes #37